+++ src/tema.cpp
@@ -146,6 +146,7 @@ void cargar(const std::wstring& carpeta_exe) {
     a.segundos_aviso = (int)j["segundos_aviso"].num(6);
     a.volumen_video = (int)j["volumen_video"].num(100);
     a.velocidad = j["velocidad"].num(1.0);
+    a.idioma_traduccion = ancho(j["idioma_traduccion"].str("English"));
     a.ventana_x = (int)j["ventana_x"].num(0);
     a.ventana_y = (int)j["ventana_y"].num(0);
     a.ventana_w = (int)j["ventana_w"].num(0);
@@ -190,6 +191,7 @@ void guardar() {
     numero("segundos_aviso", g_ajustes.segundos_aviso);
     numero("volumen_video", g_ajustes.volumen_video);
     numero("velocidad", g_ajustes.velocidad);
+    texto("idioma_traduccion", angosto(g_ajustes.idioma_traduccion));
     numero("ventana_x", g_ajustes.ventana_x);
     numero("ventana_y", g_ajustes.ventana_y);
     numero("ventana_w", g_ajustes.ventana_w);